+++ src/types.rs
@@ -4,7 +4,7 @@ use std::iter::*;
 
 
 
-pub const MAX_PLY: u32 = 128;
+pub const MAX_PLY: usize = 128;
 
 pub struct Chooser<T> {
     pub black: T,
